refactor(blog): extract route components
